Add files via upload
About cluster_centerpoints.csv

The cluster_centerpoints.csv file is a key output of the 311-Analysis Spatial Clustering Program. This CSV (Comma-Separated Values) file contains the calculated center points of spatial clusters identified from the analysis of 311 service request data. Each row in this file represents a unique cluster, providing insights into the geographic concentration of various types of service requests in the city.
File Structure

The file comprises several columns, each containing specific information about the clusters:

    ClusterID: A unique identifier assigned to each cluster.
    Year: The year in which the service requests were made.
    Week: The week of the year when the service requests were recorded.
    SR TYPE: The type of service request (e.g., Air Pollution, Potholes).
    Latitude: The latitude coordinate of the cluster's center.
    Longitude: The longitude coordinate of the cluster's center.

Usage

This file is instrumental for city planners and analysts in:

    Identifying hotspots for specific types of service requests.
    Understanding spatial patterns in urban service needs.
    Informing targeted responses and resource allocation.

The data in cluster_centerpoints.csv also serves as the basis for further spatiotemporal correlation analysis in the subsequent phase of the 311-Analysis project.